Refactor modal responsibilities
Form is now responsible for:
  - initiating/cancelling the request
  - displaying toasts on success/failure
  - resetting the file picker on completion

The modal now passes a callback to be invoked when the form upload
succeeds. This is less flexible than all state being managed in the
parent, but we can move those back up as necessary.

This also swaps implementations of useLists: we previously had a general
hook but this was cleaned up significantly through the use of react-use
hooks.