Move detections metrics schema into separate file
This was getting unwieldy, moving it out helps understand the collector
better, at least.

The @ts-expect-errors are interesting; we basically got away without
specifying these keys even though they were defined in our types because
we were passing `{}` as the type definition to the usageCollector
itself. This at least makes that glossing-over explicit.