Updates roles/tests/test roles to use new Features
This updates prebuilt roles and those used in tests to:

* Reference siemv5 instead of the older siemv4
* Add the new `rules:read` or `rules:all` feature where appropriate

Without this change, our tests are implicitly testing the
`replacedBy`/"migration" path that existing users will follow. With that
version of the code being green, we can have confidence in the behavior
for existing users, and can then update our tests to use the latest
features, here.