`buildBulkResponse` uses the statusCode of the highest found error
Instead of always returning a 500 when there are errors, we can instead
inspect the statusCode of the error (which is often present in the API
middleware).

This refactors the function responsible for building that response to
extract out a shared normalization function, and add a simple `Math.max`
call to get what we need. Since we are only dealing with 4xx and 5xx
errors in this case, we can say that 4xxs are "lower" and also default
to a 500 if we cannot determine the suggested code.