Return normalized errors from bulkGapFillByRuleIds
Simplifies the errors returned by `bulkGapFillByRuleIds` to be an
extension of `BulkOperationError`. The errors returned were previously
typed as (simplified) `BulkOperationError | any`, and we were
transforming our internal `BulkGapFillingErroredRule` manually into a
`BulkOperationError`, and losing a lot of information in the process.
Instead, we now return an extended `BulkOperationError` from the gap
filling code, and correctly capture those errors' `status` in
`buildBulkActionsResponse`.

Unit tests were also added for the interaction of
`buildBulkActionsResponse` and various `BulkOperationError` errors.

At this point, the Bulk Actions RBAC API tests are green.